location: Add source changes to enable location modules in OpenWRT.
Add location_api_msg_proto/protobuf_gen.sh script and
explicit type cast in client_api/src/LocationClientApiImpl.cpp
to enable location-api-msg-proto and location-client-api modules
for OpenWrt.

CRs-Fixed: 3217189

+++ client_api/src/LocationClientApiImpl.cpp
@@ -1365,7 +1365,7 @@ LocationClientApiImpl::LocationClientApiImpl(CapabilitiesCb capabitiescb) :
     }
 
     SockNodeLocal sock(LOCATION_CLIENT_API, pid, mClientId);
-    size_t pathNameLength = strlcpy(mSocketName, sock.getNodePathname().c_str(),
+    size_t pathNameLength = (size_t) strlcpy(mSocketName, sock.getNodePathname().c_str(),
                                     sizeof(mSocketName));
     if (pathNameLength >= sizeof(mSocketName)) {
         LOC_LOGe("socket name length exceeds limit of %" PRIu32 " bytes",
